// frontend/.schema-check-tmp/models/ArchiveShotResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ArchiveShotResponse = {
  projectId: string;
  shotId: string;
  fileName: string;
  fileSize: number;
  fileCount: number;
  downloadUrl: string;
  layerCount?: number;
  meshCount?: number;
};