// frontend/.schema-check-tmp/models/CreateShotRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateShotRequest = {
  /**
   * Shot ID
   */
  shotId: string;
  /**
   * Shot description
   */
  description?: (string | null);
  /**
   * Scene type (e.g. indoor, outdoor)
   */
  sceneType?: (string | null);
};